+++ core/editor/.eslintrc.js
@@ -13,6 +13,8 @@ module.exports = {
     rules: {
         "@typescript-eslint/no-explicit-any": 1,
         "@typescript-eslint/no-var-requires": 1,
-        "@typescript-eslint/ban-types": 1
+        "@typescript-eslint/ban-types": 1,
+        "@typescript-eslint/no-unused-vars": 1,
+        "@typescript-eslint/no-empty-function": 1
     }
 };