+++ core/editor/src/ts/sections/backgroundMask.ts
@@ -1,9 +1,8 @@
-import { GUI } from "dat.gui";
 import { Container } from "tsparticles/dist/Core/Container";
 import { changeHandler } from "../utils";
 
-const addBackgroundMask = (gui: GUI, container: Container): void => {
-    const fBgMask = gui.addFolder("backgroundMask");
+const addBackgroundMask = (container: Container): void => {
+    /*const fBgMask = addFolder("backgroundMask");
     const fCover = fBgMask.addFolder("cover");
     const fCoverColor = fBgMask.addFolder("color");
 
@@ -14,6 +13,7 @@ const addBackgroundMask = (gui: GUI, container: Container): void => {
     fCover.add(container.options.backgroundMask.cover, "opacity").onChange(async () => changeHandler(container));
 
     fBgMask.add(container.options.backgroundMask, "enable").onChange(async () => changeHandler(container));
+     */
 };
 
 export { addBackgroundMask };