+++ core/tsparticles/src/Core/Interfaces/IMouseData.ts
@@ -3,6 +3,7 @@ import type { ICoordinates } from "./ICoordinates";
 export interface IMouseData {
     clickPosition?: ICoordinates;
     position?: ICoordinates;
+    downPosition?: ICoordinates;
     clickTime?: number;
     inside: boolean;
     clicking: boolean;