+++ demo/main/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.8](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.2.0-alpha.7...tsparticles-demo@1.2.0-alpha.8) (2020-06-26)
+
+
+### Bug Fixes
+
+* click handler with particle size gives better results, little fixes for absorber orbits ([e851049](https://github.com/matteobruni/tsparticles/commit/e851049b480a205813b9df786ad7c5761645de1c))
+
+
+
+
+
+# [1.2.0-alpha.7](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.2.0-alpha.6...tsparticles-demo@1.2.0-alpha.7) (2020-06-26)
+
+**Note:** Version bump only for package tsparticles-demo
+
+
+
+
+
+# [1.2.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.2.0-alpha.5...tsparticles-demo@1.2.0-alpha.6) (2020-06-24)
+
+
+### Features
+
+* working on orbits for absorbers ([1f55430](https://github.com/matteobruni/tsparticles/commit/1f55430b51a108236940e6a8c3d2ae97c82583b2))
+
+
+
+
+
+# [1.2.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.1.1...tsparticles-demo@1.2.0-alpha.5) (2020-06-23)
+
+
+### Bug Fixes
+
+* forwad demo improved ([3bb0f74](https://github.com/matteobruni/tsparticles/commit/3bb0f7406f2cea92c288b3fc092c002f9e465620))
+* updated index.html ([bd3dc5d](https://github.com/matteobruni/tsparticles/commit/bd3dc5dab9676d58b749a28af2436cf04707e1d1))
+
+
+### Features
+
+* added option for rotate following move direction, closes [#408](https://github.com/matteobruni/tsparticles/issues/408) ([ee1dd85](https://github.com/matteobruni/tsparticles/commit/ee1dd850fafd1e00e750c9738f4f60543133c504))
+* draggable option for absorbers ([3b42861](https://github.com/matteobruni/tsparticles/commit/3b4286126800c316768326f774f13aa2cda62198))
+* mouse trail completed, closes [#401](https://github.com/matteobruni/tsparticles/issues/401) ([4b1bc1a](https://github.com/matteobruni/tsparticles/commit/4b1bc1a185860718e26904c6980f88621d4a99ce))
+* particles orientation based on their direction, closes [#408](https://github.com/matteobruni/tsparticles/issues/408) ([f0ec58e](https://github.com/matteobruni/tsparticles/commit/f0ec58ebf463514805272871fb3c856e3d667aee))
+
+
+
+
+
+# [1.2.0-alpha.4](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.1.0...tsparticles-demo@1.2.0-alpha.4) (2020-06-22)
+
+
+### Bug Fixes
+
+* forwad demo improved ([3bb0f74](https://github.com/matteobruni/tsparticles/commit/3bb0f7406f2cea92c288b3fc092c002f9e465620))
+* updated index.html ([bd3dc5d](https://github.com/matteobruni/tsparticles/commit/bd3dc5dab9676d58b749a28af2436cf04707e1d1))
+
+
+### Features
+
+* added option for rotate following move direction, closes [#408](https://github.com/matteobruni/tsparticles/issues/408) ([ee1dd85](https://github.com/matteobruni/tsparticles/commit/ee1dd850fafd1e00e750c9738f4f60543133c504))
+* draggable option for absorbers ([3b42861](https://github.com/matteobruni/tsparticles/commit/3b4286126800c316768326f774f13aa2cda62198))
+* mouse trail completed, closes [#401](https://github.com/matteobruni/tsparticles/issues/401) ([4b1bc1a](https://github.com/matteobruni/tsparticles/commit/4b1bc1a185860718e26904c6980f88621d4a99ce))
+* particles orientation based on their direction, closes [#408](https://github.com/matteobruni/tsparticles/issues/408) ([f0ec58e](https://github.com/matteobruni/tsparticles/commit/f0ec58ebf463514805272871fb3c856e3d667aee))
+
+
+
+
+
 # [1.2.0-alpha.3](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.1.0...tsparticles-demo@1.2.0-alpha.3) (2020-06-21)