+++ shapes/bubble/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.8](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.2.0-alpha.7...tsparticles-shape-bubble@1.2.0-alpha.8) (2020-06-26)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
+# [1.2.0-alpha.7](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.2.0-alpha.6...tsparticles-shape-bubble@1.2.0-alpha.7) (2020-06-26)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
+# [1.2.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.2.0-alpha.5...tsparticles-shape-bubble@1.2.0-alpha.6) (2020-06-24)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
+# [1.2.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.1.1...tsparticles-shape-bubble@1.2.0-alpha.5) (2020-06-23)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
+# [1.2.0-alpha.4](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.1.0...tsparticles-shape-bubble@1.2.0-alpha.4) (2020-06-22)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
 # [1.2.0-alpha.3](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.1.0...tsparticles-shape-bubble@1.2.0-alpha.3) (2020-06-21)
 
 **Note:** Version bump only for package tsparticles-shape-bubble