+++ wrappers/angular/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.17.0-alpha.8](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.17.0-alpha.7...ng-particles@1.17.0-alpha.8) (2020-06-26)
+
+**Note:** Version bump only for package ng-particles
+
+
+
+
+
+# [1.17.0-alpha.7](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.17.0-alpha.6...ng-particles@1.17.0-alpha.7) (2020-06-26)
+
+**Note:** Version bump only for package ng-particles
+
+
+
+
+
+# [1.17.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.17.0-alpha.5...ng-particles@1.17.0-alpha.6) (2020-06-24)
+
+**Note:** Version bump only for package ng-particles
+
+
+
+
+
+# [1.17.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.16.1...ng-particles@1.17.0-alpha.5) (2020-06-23)
+
+
+### Bug Fixes
+
+* angular build fixed ([749666c](https://github.com/matteobruni/tsparticles/commit/749666c1e0fee4cbe5667bf4e6f5eaab52d1e5fd))
+
+
+
+
+
+# [1.17.0-alpha.4](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.16.0...ng-particles@1.17.0-alpha.4) (2020-06-22)
+
+
+### Bug Fixes
+
+* angular build fixed ([749666c](https://github.com/matteobruni/tsparticles/commit/749666c1e0fee4cbe5667bf4e6f5eaab52d1e5fd))
+
+
+
+
+
 # [1.17.0-alpha.3](https://github.com/matteobruni/tsparticles/compare/ng-particles@1.16.0...ng-particles@1.17.0-alpha.3) (2020-06-21)