+++ wrappers/vue/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.17.0-alpha.8](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.17.0-alpha.7...particles.vue@1.17.0-alpha.8) (2020-06-26)
+
+**Note:** Version bump only for package particles.vue
+
+
+
+
+
+# [1.17.0-alpha.7](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.17.0-alpha.6...particles.vue@1.17.0-alpha.7) (2020-06-26)
+
+**Note:** Version bump only for package particles.vue
+
+
+
+
+
+# [1.17.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.17.0-alpha.5...particles.vue@1.17.0-alpha.6) (2020-06-24)
+
+**Note:** Version bump only for package particles.vue
+
+
+
+
+
+# [1.17.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.16.1...particles.vue@1.17.0-alpha.5) (2020-06-23)
+
+**Note:** Version bump only for package particles.vue
+
+
+
+
+
+# [1.17.0-alpha.4](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.16.0...particles.vue@1.17.0-alpha.4) (2020-06-22)
+
+**Note:** Version bump only for package particles.vue
+
+
+
+
+
 # [1.17.0-alpha.3](https://github.com/matteobruni/tsparticles/compare/particles.vue@1.16.0...particles.vue@1.17.0-alpha.3) (2020-06-21)
 
 **Note:** Version bump only for package particles.vue